% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_builder.R
\name{add_body}
\alias{add_body}
\title{Add a body to a response}
\usage{
add_body(res, body)
}
\arguments{
\item{res}{A response builder object}

\item{body}{The body to add}
}
\value{
The response builder object
}
\description{
Adds a body to a response, if no content type is set, it will be detected
}
