% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_builder.R
\name{add_body_json}
\alias{add_body_json}
\title{Add a body to a response as JSON}
\usage{
add_body_json(res, body)
}
\arguments{
\item{res}{A response builder object}

\item{body}{The body to add}
}
\value{
The response builder object
}
\description{
Adds a body to a response as JSON
}
