% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_builder.R
\name{add_cookie}
\alias{add_cookie}
\title{Add a cookie to a response}
\usage{
add_cookie(res, name, value)
}
\arguments{
\item{res}{A response builder object}

\item{name}{The name of the cookie}

\item{value}{The value of the cookie}
}
\value{
The response builder object
}
\description{
Adds a cookie to a response
}
