% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{add_delete_route}
\alias{add_delete_route}
\title{Add a DELETE route}
\usage{
add_delete_route(tower, path, handler)
}
\arguments{
\item{tower}{A tower object}

\item{path}{A string containing the path to match}

\item{handler}{A function to call when the route is matched}
}
\value{
The tower with the added DELETE route
}
\description{
Adds a DELETE route to a tower
}
