% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{add_http_layer}
\alias{add_http_layer}
\title{Add an HTTP layer to a tower}
\usage{
add_http_layer(tower, layer)
}
\arguments{
\item{tower}{A tower}

\item{layer}{A function that takes a request and returns either
a response or NULL. NULL indicates that the layer did not
short-circuit the request, therefore the next layer should be
called. If a response is returned, the request is short-circuited
and the response is returned to the client.}
}
\value{
The tower with the layer added
}
\description{
Add an HTTP layer to a tower. This layer
will be called before the 'shiny' app's httpHandler.
}
