% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-add-metadata.R
\name{add_linestarts}
\alias{add_linestarts}
\title{Add a default line-start value to the scaffold data-frame
in preparation for loading the data into R}
\usage{
add_linestarts(scaffold_df, default = 1)
}
\arguments{
\item{scaffold_df}{A tibble/data.frame containing columns:
\enumerate{
\item filename - Name of the file
\item filepath - Path to the file
\item file_ext - File extension (e.g. .csv)
}

And any other columns required by the function.}

\item{default}{Default value to use (added to all rows)}
}
\value{
\code{scaffold_df} with an additional column indicating
the line to start reading the data from
}
\description{
Add a default line-start value to the scaffold data-frame
in preparation for loading the data into R
}
\examples{

library(magrittr)

tibble::tibble(
  filename = c("a_101.csv", "a_201.csv", "b_201.csv"),
  filepath = c("data/a_101.csv", "data/a_201.csv", "data/b_201.csv"),
  file_ext = c("csv", "csv", "csv")
) \%>\%
 metadata_col_from_regex(
   scaffold_df = .,
   regex_list = list(
     "seminar_series" = "\\\\d\\\\d\\\\d"
   )
 ) \%>\%
 add_linestarts(default = 1)
}
