% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{add_patch_route}
\alias{add_patch_route}
\title{Add a PATCH route}
\usage{
add_patch_route(tower, path, handler)
}
\arguments{
\item{tower}{A tower object}

\item{path}{A string containing the path to match}

\item{handler}{A function to call when the route is matched}
}
\value{
The tower with the added PATCH route
}
\description{
Adds a PATCH route to a tower
}
