% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{add_post_route}
\alias{add_post_route}
\title{Add a POST route}
\usage{
add_post_route(tower, path, handler)
}
\arguments{
\item{tower}{A tower object}

\item{path}{A string containing the path to match}

\item{handler}{A function to call when the route is matched}
}
\value{
The tower with the added POST route
}
\description{
Adds a POST route to a tower
}
