% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{add_put_route}
\alias{add_put_route}
\title{Add a PUT route}
\usage{
add_put_route(tower, path, handler)
}
\arguments{
\item{tower}{A tower object}

\item{path}{A string containing the path to match}

\item{handler}{A function to call when the route is matched}
}
\value{
The tower with the added PUT route
}
\description{
Adds a PUT route to a tower
}
