% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{add_route}
\alias{add_route}
\title{Add an HTTP layer to a tower}
\usage{
add_route(tower, method = "GET", path, handler)
}
\arguments{
\item{tower}{A tower object}

\item{method}{A string containing the HTTP method to match}

\item{path}{A string containing the path to match}

\item{handler}{A function to call when the layer is matched}
}
\value{
The tower with the added route
}
\description{
Adds an HTTP layer to a tower
}
