% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{add_server_layer}
\alias{add_server_layer}
\title{Add a server layer to a tower}
\usage{
add_server_layer(tower, layer)
}
\arguments{
\item{tower}{A tower}

\item{layer}{A function that takes input, output, and session
and has no return value. This function will be called before
the original server function. If you want to short-circuit
the server use an exception.}
}
\value{
The tower with the added layer
}
\description{
Add a server layer to a tower. This layer
will run before the 'shiny' app's server function. This
is useful for adding custom logic to the server function
without modifying the original server function.
}
