% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-add-metadata.R
\name{add_sheet_selections}
\alias{add_sheet_selections}
\title{Add a default name or index of the sheet to read for each file}
\usage{
add_sheet_selections(scaffold_df, default = 1)
}
\arguments{
\item{scaffold_df}{A tibble/data.frame containing columns:
\enumerate{
\item filename - Name of the file
\item filepath - Path to the file
\item file_ext - File extension (e.g. .csv)
}

And any other columns required by the function.}

\item{default}{Default value to use (added to all rows)}
}
\value{
\code{scaffold_df} with a \code{sheet_selection} column added, containing
the default value supplied.
}
\description{
Note: this only affects the reading of excel/spreadsheet files (e.g. workbooks
containing multiple sheets)
}
