% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{app_into_parts}
\alias{app_into_parts}
\title{Into parts}
\usage{
app_into_parts(app)
}
\arguments{
\item{app}{A shiny.appobj}
}
\value{
A list with the ui and server handlers
}
\description{
Splits a shiny.appobj into its parts, the ui and server
}
