% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-vertical-merge.R
\name{apply_selections}
\alias{apply_selections}
\title{Extract selected columns from datasets in scaffold}
\usage{
apply_selections(scaffold_df, selections)
}
\arguments{
\item{scaffold_df}{scaffold dataframe}

\item{selections}{TRUE/character vector, keep all (TRUE) or selected
(colnames in character vector) columns in each dataframe in the scaffold}
}
\value{
\code{scaffold_df} with column selections applied to the datasets
}
\description{
Extract selected columns from datasets in scaffold
}
