% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{build_diagnostic_quick_summary}
\alias{build_diagnostic_quick_summary}
\title{build_diagnostic_quick_summary}
\usage{
build_diagnostic_quick_summary(diagnostics_list, tower_df)
}
\arguments{
\item{diagnostics_list}{A list of dfs containing diagnostic summaries of
for different aspects of the merge}

\item{tower_df}{The final merged dataset}
}
\value{
string message
}
\description{
build_diagnostic_quick_summary
}
