% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{build_http_cookie}
\alias{build_http_cookie}
\title{Build a cookie}
\usage{
build_http_cookie(key, value)
}
\arguments{
\item{key}{A string containing the cookie key}

\item{value}{A string containing the cookie value}
}
\value{
A string containing the formated cookie
}
\description{
Builds an HttpOnly cookie from a key and value
}
