% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_builder.R
\name{build_response}
\alias{build_response}
\title{Build a response}
\usage{
build_response(res)
}
\arguments{
\item{res}{A response builder object}
}
\value{
A 'shiny' response object
}
\description{
Builds a response
}
