% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{build_tower}
\alias{build_tower}
\title{Build a Shiny app from a tower}
\usage{
build_tower(tower)
}
\arguments{
\item{tower}{A tower}
}
\value{
A Shiny app object
}
\description{
Build a Shiny app from a tower. This will create
a new Shiny app with the specified layers added.
}
