% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-file-coverage.R
\name{check_metadata_coverage}
\alias{check_metadata_coverage}
\title{check_metadata_coverage}
\usage{
check_metadata_coverage(scaffold_df, meta_col, expected_values)
}
\arguments{
\item{expected_values}{}
}
\value{

}
\description{
check_metadata_coverage
}
