% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{classify_column_basic}
\alias{classify_column_basic}
\title{classify_column_basic}
\usage{
classify_column_basic(colname, df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
classify_column_basic
}
