% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_nothing.R
\name{do_nothing}
\alias{do_nothing}
\title{Do nothing}
\usage{
do_nothing(data, unused = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{unused}{Unused.}
}
\value{
\code{data}
}
\description{
Do nothing
}
\examples{
data <- BOD
do_nothing(data)
}
