% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{eval_missingness_diagnostics}
\alias{eval_missingness_diagnostics}
\title{eval_missingness_diagnostics}
\usage{
eval_missingness_diagnostics(diag_ob, tower_df)
}
\arguments{
\item{diag_ob}{The diagnostic df}

\item{tower_df}{The final merged data}
}
\value{
boolean
}
\description{
eval_missingness_diagnostics
}
