% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{get_tower_diagnostics}
\alias{get_tower_diagnostics}
\title{get_tower_diagnostics}
\usage{
get_tower_diagnostics(df)
}
\arguments{
\item{df}{A df produced by a call to toweR::vertical_merge(scaffold_df, keep_diagnostics = TRUE)}
}
\value{
The a list of dataframes providing diagnostic checks for the vertical
merge used to combine the different datasets.
}
\description{
Get the diagnostics from a dataframe constructed
using toweR::vertical_merge(scaffold_df, keep_diagnostics = TRUE)
}
\examples{

\dontrun{
 final_df <- vertical_merge(scaffold_df, keep_diagnostics = TRUE)
 full_merge_diagnostics <- get_tower_diagnostics(final_df)
}

}
