% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-harmonise-data.R
\name{harmonise_by_explicit}
\alias{harmonise_by_explicit}
\title{harmonise_by_explicit}
\usage{
harmonise_by_explicit(scaffold_df, capture_list, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
harmonise_by_explicit
}
