% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-vertical-merge.R
\name{include_metadata}
\alias{include_metadata}
\title{include_metadata}
\usage{
include_metadata(scaffold_df, add_metadata)
}
\arguments{
\item{scaffold_df}{scaffold_df}

\item{add_metadata}{boolean}
}
\value{
\code{scaffold_df} with metadata
}
\description{
include_metadata
}
