% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-runtime-validation.R
\name{is_file_scaffold}
\alias{is_file_scaffold}
\title{Standardised runtime assertion: is file scaffold}
\usage{
is_file_scaffold(ob)
}
\arguments{
\item{ob}{A candidate scaffold_df object}
}
\value{
TRUE
}
\description{
Standardised runtime assertion: is file scaffold
}
