% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-add-metadata.R
\name{modify_linestarts}
\alias{modify_linestarts}
\title{Add a default line-start value to the scaffold data-frame
in preparation for loading the data into R}
\usage{
modify_linestarts(
  scaffold_df,
  meta_col = "filename",
  meta_values,
  new_linestart
)
}
\arguments{
\item{scaffold_df}{A tibble/data.frame containing columns:
\enumerate{
\item filename - Name of the file
\item filepath - Path to the file
\item file_ext - File extension (e.g. .csv)
}}

\item{meta_col}{A string, the name of the meta column to use to identify
files whose meta-data should be changed}

\item{meta_values}{Rows matching these values will replace their linestart value with \code{new_linestart}}

\item{new_linestart}{A number or string}
}
\value{

}
\description{
Add a default line-start value to the scaffold data-frame
in preparation for loading the data into R
}
