% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-add-metadata.R
\name{modify_sheet_selections}
\alias{modify_sheet_selections}
\title{Add a default line-start value to the scaffold data-frame
in preparation for loading the data into R}
\usage{
modify_sheet_selections(
  scaffold_df,
  meta_col = "filename",
  meta_values,
  new_sheets_id
)
}
\arguments{
\item{default}{}
}
\value{

}
\description{
Add a default line-start value to the scaffold data-frame
in preparation for loading the data into R
}
