% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{print.tower}
\alias{print.tower}
\title{Print a tower}
\usage{
\method{print}{tower}(x, ...)
}
\arguments{
\item{x}{A tower}

\item{...}{Ignored arguments (for compatibility with print)}
}
\value{
No return value, called for side effects
}
\description{
Print a tower
}
