% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-load-data.R
\name{read_helper_csv}
\alias{read_helper_csv}
\title{read_helper_csv}
\usage{
read_helper_csv(scaffold_row)
}
\arguments{
\item{scaffold_row}{A row of a \code{scaffold_df} dataframe}
}
\value{
A dataframe
}
\description{
read_helper_csv
}
