% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{req_body_form}
\alias{req_body_form}
\title{Extract form data from a request}
\usage{
req_body_form(req)
}
\arguments{
\item{req}{A request object}
}
\value{
A list containing the form data in the body
}
\description{
Extracts form data from a request
}
