% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{req_cookies}
\alias{req_cookies}
\title{Extract cookies from a request}
\usage{
req_cookies(req)
}
\arguments{
\item{req}{A request object}
}
\value{
A list containing the cookies
}
\description{
Extracts cookies from a request
}
