% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{req_query}
\alias{req_query}
\title{Extract query parameters from a request}
\usage{
req_query(req)
}
\arguments{
\item{req}{A request object}
}
\value{
A list containing the query parameters
}
\description{
Extracts query parameters from a request
}
