% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{run_coltype_diagnostics}
\alias{run_coltype_diagnostics}
\title{run_coltype_diagnostics}
\usage{
run_coltype_diagnostics(scaffold_df, tower_df)
}
\arguments{
\item{scaffold_df}{scaffold df containing datasets to be merged}

\item{tower_df}{A single df containing the vertically merged data}
}
\value{
diagnostic df
}
\description{
run_coltype_diagnostics
}
