% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{run_diagnostics}
\alias{run_diagnostics}
\title{run_diagnostics}
\usage{
run_diagnostics(scaffold_df, tower_df, diagnostics)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
run_diagnostics
}
