% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-metadata.R
\name{safe_left_join}
\alias{safe_left_join}
\title{Perform left join defensively}
\usage{
safe_left_join(df_l, df_r, by)
}
\arguments{
\item{df_l}{Left-hand df}

\item{df_r}{Right hand df}

\item{by}{String, name of the key column}
}
\value{
A df of \code{df_r} left joined into \code{df_l}
}
\description{
Will fail if:
}
\details{
\enumerate{
\item Neither of the dfs have a unique \code{by} key
\item The rowcount of the output df is not equal to
the rowcount of the left-hand df
}
}
