% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_builder.R
\name{set_content_type}
\alias{set_content_type}
\title{Set the content type of a response}
\usage{
set_content_type(res, content_type)
}
\arguments{
\item{res}{A response builder object}

\item{content_type}{The content type to set}
}
\value{
The response builder object
}
\description{
Sets the content type of a response
}
