% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_builder.R
\name{set_header}
\alias{set_header}
\title{Set a header on a response}
\usage{
set_header(res, name, value)
}
\arguments{
\item{res}{A response builder object}

\item{name}{The name of the header}

\item{value}{The value of the header}
}
\value{
The response builder object
}
\description{
Sets or adds a header to a response
}
