% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_builder.R
\name{set_status}
\alias{set_status}
\title{Set the status of a response}
\usage{
set_status(res, status)
}
\arguments{
\item{res}{A response builder object}

\item{status}{The status to set}
}
\value{
The response builder object
}
\description{
Sets the status of a response
}
