% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-load-data.R
\name{simple_read_switch}
\alias{simple_read_switch}
\title{simple_read_switch}
\usage{
simple_read_switch(scaffold_row)
}
\arguments{
\item{scaffold_row}{}
}
\value{

}
\description{
simple_read_switch
}
