% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-other.R
\name{string_alphanum_only}
\alias{string_alphanum_only}
\title{Remove any non-alphanumeric characters from a character vector}
\usage{
string_alphanum_only(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
x with all instances of non-alphanumeric character replaced
with an empty character
}
\description{
Remove any non-alphanumeric characters from a character vector
}
