% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{add_http_layer}
\alias{add_http_layer}
\title{Add an HTTP layer to a tower}
\usage{
add_http_layer(tower, layer)
}
\arguments{
\item{tower}{A tower}

\item{layer}{A function that takes a request and returns either
a response. A layer can short circuit by returning a response
directly or call the next layer will \code{req$NEXT(req)} which
will call the next layer in the middleware.}
}
\value{
The tower with the added layer
}
\description{
Add an HTTP layer to a tower. This layer
will be called before the 'shiny' app's httpHandler.
}
