% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-add-metadata.R
\name{add_linestarts}
\alias{add_linestarts}
\title{Add a default line-start value to the scaffold data-frame
in preparation for loading the data into R}
\usage{
add_linestarts(scaffold_df, default = 1)
}
\arguments{
\item{scaffold_df}{A tibble/data.frame containing columns:
\enumerate{
\item filename - Name of the file
\item filepath - Path to the file
\item file_ext - File extension (e.g. .csv)
}}

\item{default}{A number, the row to begin reading the file from}
}
\value{

}
\description{
Add a default line-start value to the scaffold data-frame
in preparation for loading the data into R
}
