% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-add-metadata.R
\name{add_sheet_selections}
\alias{add_sheet_selections}
\title{Add a default line-start value to the scaffold data-frame}
\usage{
add_sheet_selections(scaffold_df, default = 1)
}
\arguments{
\item{scaffold_df}{A tibble/data.frame containing columns:
\enumerate{
\item filename - Name of the file
\item filepath - Path to the file
\item file_ext - File extension (e.g. .csv)
}}

\item{default}{A number or string}

\item{load_all_sheets}{}
}
\value{

}
\description{
Add a default line-start value to the scaffold data-frame
}
