% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{build_diagnostic_quick_summary}
\alias{build_diagnostic_quick_summary}
\title{build_diagnostic_quick_summary}
\usage{
build_diagnostic_quick_summary(diagnostics_list, tower_df)
}
\arguments{
\item{tower_df}{}
}
\value{

}
\description{
build_diagnostic_quick_summary
}
