% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-scaffold.R
\name{build_scaffolding_from_dir}
\alias{build_scaffolding_from_dir}
\title{Build a data scaffold (tibble) from the contents of a directory}
\usage{
build_scaffolding_from_dir(dir_path, file_exts = NULL, verbose = FALSE)
}
\arguments{
\item{dir_path}{File directory path (string)}

\item{file_exts}{If NULL, all files in the directory as included in the scaffold,
otherwise, if a character vector is supplied, all files with matching extensions (e.g. 'csv')
are included. Any non-alphanumeric characters will be removed (e.g. '.csv' will become 'csv')
to match with the extracted file extension.}

\item{verbose}{If TRUE, prints list of the files added}
}
\value{
A data.frame/tibble with one row per-file, and columns with file metadata:
\enumerate{
\item filename - Name of the file
\item filepath - Path to the file
\item file_ext - File extension (e.g. .csv)
}
}
\description{
Build a data scaffold (tibble) from the contents of a directory
}
