% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{build_tower}
\alias{build_tower}
\title{Build a 'shiny' app from a tower}
\usage{
build_tower(tower)
}
\arguments{
\item{tower}{A tower}
}
\value{
A 'shiny' app object that can be started
}
\description{
Build a 'shiny' app from a tower. This will create
a new 'shiny' app with the specified layers added.
}
