% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{create_tower}
\alias{create_tower}
\title{Create a new tower}
\usage{
create_tower(app)
}
\arguments{
\item{app}{A 'shiny' app object}
}
\value{
A new tower object to add more layers to
}
\description{
Create a new tower to build upon.
}
