% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{eval_coltype_diagnostics}
\alias{eval_coltype_diagnostics}
\title{eval_coltype_diagnostics}
\usage{
eval_coltype_diagnostics(diag_ob, tower_df)
}
\arguments{
\item{tower_df}{}
}
\value{

}
\description{
eval_coltype_diagnostics
}
