% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-load-data.R
\name{load_materials_simple}
\alias{load_materials_simple}
\title{load_materials_simple}
\usage{
load_materials_simple(scaffold_df)
}
\arguments{
\item{scaffold_df}{}
}
\value{

}
\description{
load_materials_simple
}
