% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-add-metadata.R
\name{metadata_col_from_regex}
\alias{metadata_col_from_regex}
\title{Add metadata columns to a scaffold using regex to
extract information from the filename}
\usage{
metadata_col_from_regex(scaffold_df, regex_list)
}
\arguments{
\item{scaffold_df}{A tibble/data.frame containing columns:
\enumerate{
\item filename - Name of the file
\item filepath - Path to the file
\item file_ext - File extension (e.g. .csv)
}}

\item{regex_list}{A named list of strings, with the names corresponding
to the new columns to be created, and the strings representing
regex patterns used to extract information from the filename column.}
}
\value{
\code{scaffold_df} with additional columns created by applying \code{regex_list}
}
\description{
Add metadata columns to a scaffold using regex to
extract information from the filename
}
