% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{missingness_col_basic}
\alias{missingness_col_basic}
\title{missingness_col_basic}
\usage{
missingness_col_basic(colname, df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
missingness_col_basic
}
