% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-harmonise-data.R
\name{rename_cols_by_capture_list}
\alias{rename_cols_by_capture_list}
\title{rename_cols_by_capture_list}
\usage{
rename_cols_by_capture_list(df, df_name, capture_list, verbose)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
rename_cols_by_capture_list
}
