% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{run_colname_diagnostics}
\alias{run_colname_diagnostics}
\title{run_colname_diagnostics}
\usage{
run_colname_diagnostics(scaffold_df, tower_df)
}
\arguments{
\item{scaffold_df}{scaffold df containing datasets to be merged}

\item{tower_df}{A single df containing the vertically merged data}
}
\value{
diagnostics_df
}
\description{
run_colname_diagnostics
}
