% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-merge-diagnostics.R
\name{run_coltype_diagnostics}
\alias{run_coltype_diagnostics}
\title{run_coltype_diagnostics}
\usage{
run_coltype_diagnostics(scaffold_df, tower_df)
}
\arguments{
\item{tower_df}{}
}
\value{

}
\description{
run_coltype_diagnostics
}
